add: add backend check to connection check and dialog
