+++ src/RouterSelector.web.js
@@ -1,7 +1,6 @@
 import React, { useEffect } from 'react'
 import { AsyncStorage } from 'react-native'
 import bip39 from 'bip39-light'
-import API from './lib/API/api'
 import { DESTINATION_PATH } from './lib/constants/localStorage'
 import SimpleStore from './lib/undux/SimpleStore'
 import Splash from './components/splash/Splash'
@@ -14,42 +13,11 @@ import Config from './config/config'
 const log = logger.child({ from: 'RouterSelector' })
 log.debug({ Config })
 
-/**
- * Don't start app if server isn't responding
- */
-const apiReady = async () => {
-  try {
-    await API.ready
-    const res = await Promise.race([
-      API.auth()
-        .then(_ => true)
-        .catch(_ => _.message !== 'Network Error'),
-      delay(3000).then(_ => 'timeout'),
-    ])
-    log.debug({ res })
-    if (res !== true) {
-      await delay(3000)
-      return apiReady()
-    }
-    return
-  } catch (e) {
-    log.debug('apiReady:', e.message)
-    await delay(3000)
-
-    // return apiReady()
-  }
-}
-
 // import Router from './SignupRouter'
 let SignupRouter = React.lazy(() =>
   initAnalytics()
     .then(_ =>
-      Promise.all([
-        import(/* webpackChunkName: "signuprouter" */ './SignupRouter'),
-        apiReady(),
-        recoverByMagicLink(),
-        delay(2000),
-      ])
+      Promise.all([import(/* webpackChunkName: "signuprouter" */ './SignupRouter'), recoverByMagicLink(), delay(2000)])
     )
     .then(r => r[0])
 )
@@ -93,12 +61,12 @@ let AppRouter = React.lazy(() => {
   log.debug('initializing storage and wallet...')
   let walletAndStorageReady = import(/* webpackChunkName: "init" */ './init')
   let p2 = walletAndStorageReady.then(({ init, _ }) => init()).then(_ => log.debug('storage and wallet ready'))
-  return Promise.all([p2, apiReady(), import(/* webpackChunkName: "router" */ './Router')])
+  return Promise.all([import(/* webpackChunkName: "router" */ './Router'), p2])
     .then(r => {
       log.debug('router ready')
       return r
     })
-    .then(r => r[2])
+    .then(r => r[0])
 })
 
 const RouterSelector = () => {
